+++ lib/radius/cli/app.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'thor'
+require "radius/cli/puma_dev"
 
 module Radius
   module Cli
@@ -8,6 +9,8 @@ module Radius
       def self.exit_on_failure?
         true
       end
+
+      register ::Radius::Cli::PumaDev, "puma_dev", "puma_dev", PumaDev::DESCRIPTION
     end
   end
 end